+++ lib/phoenix_kit_web/components/admin_nav.ex
@@ -321,6 +321,13 @@ defmodule PhoenixKitWeb.Components.AdminNav do
             </li>
           <% end %>
 
+          <PhoenixKitWeb.Components.Core.RoleSwitcher.role_switcher
+            id="admin-role-switcher-menu"
+            scope={@scope}
+            current_path={@current_path}
+            current_locale={@current_locale}
+          />
+
           <%= if @multi_session_allowed? do %>
             <div class="divider my-0"></div>